1 component 1 file
